[MOVE] PhoneVerificationEntity를 entity 패키지 하위로 이동
