+++ src/main/java/sopt/makers/authentication/database/rdb/repository/auth/PhoneVerificationRetriever.java
@@ -1,6 +1,6 @@
 package sopt.makers.authentication.database.rdb.repository.auth;
 
-import sopt.makers.authentication.database.rdb.entity.auth.PhoneVerificationEntity;
+import sopt.makers.authentication.database.rdb.entity.PhoneVerificationEntity;
 import sopt.makers.authentication.domain.auth.PhoneVerification;
 import sopt.makers.authentication.support.code.domain.failure.AuthFailure;
 import sopt.makers.authentication.support.exception.domain.AuthException;
